COSC499-CapstoneProject/team-14-capstone-team-14-capstone-develop/app/backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==4.2.22
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-dotenv==1.1.0
python-docx==1.1.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
requests
django-log-request-id>=2.1.0
coverage==7.5.0
isort==5.13.2
flake8==7.0.0
django-debug-toolbar==4.2.0
Pillow
drf-nested-routers
reportlab
weasyprint
openpyxl==3.1.2
xlsxwriter==3.1.2
numpy==1.24.3
scipy==1.11.4
psutil==6.1.0